+++ lang/de/messages.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+
+    'required' => 'Bitte ausfüllen.',
+    'email' => 'Bitte eine gültige E-Mail-Adresse eingeben.',
+    'confirm' => 'Bitte bestätigen.',
+
+];